atomisp2: css2401 and 2401_legacy-irci_master_20150303_1500
Merge CSS2.1 for css2401 and 2401_legacy
for irci_master_20150303_1500 firmware tag

Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-24134
Reviewed-on: https://icggerrit.ir.intel.com/44561
Build: viefandr
Integrator: Wei Tao <wei.tao@intel.com>
Approver: Wei Tao <wei.tao@intel.com>
Build: Wei Tao <wei.tao@intel.com>
Maintainer: Wei Tao <wei.tao@intel.com>

+++ drivers/external_drivers/camera/drivers/media/pci/atomisp2/css2401a0_legacy_v21/runtime/isys/src/csi_rx_rmgr.c
@@ -44,8 +44,8 @@ bool ia_css_isys_csi_rx_lut_rmgr_acquire(
 	csi_rx_backend_lut_entry_t	*entry)
 {
 	bool retval = false;
-	uint16_t max_num_packets_of_type;
-	uint16_t num_active_of_type;
+	uint32_t max_num_packets_of_type;
+	uint32_t num_active_of_type;
 	isys_csi_rx_rsrc_t *cur_rsrc = NULL;
 	uint16_t i;
