+++ drivers/external_drivers/camera/drivers/media/pci/atomisp2/css2401a0_v21/isp/kernels/xnr/xnr_1.0/ia_css_xnr.host.c
@@ -42,9 +42,9 @@ ia_css_xnr_encode(
 	unsigned size)
 {
 	(void)size;
-	
+
 	to->threshold =
-		uDIGIT_FITTING(from->threshold, 16, SH_CSS_ISP_YUV_BITS);
+		(uint16_t)uDIGIT_FITTING(from->threshold, 16, SH_CSS_ISP_YUV_BITS);
 }
 
 void
